+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # debug-dump
 
+## 0.1.1.0
+
+  * Add `traceBSLSuffix` and `traceTextSuffix`
+
+## 0.1.0.2
+
+  * Revert unnecessary strictness
+
+## 0.1.0.1
+
+  * Bug fix for strictness
+
 ## 0.1.0.0
 
 Initial release